Update from Roman to compare the right values that get changed
Change asserts for localPosition with ones that assert when
scenePosition and globalPosition do not hold the values that were set.

Roman indicated that the local position is changed after the detach, so
that would not be shared. The scenePosition and globalPosition are the
ones that are set before the detach and that would propagate to prior
instances.

With asserts on the scenePosition and globalPosition, the asserts
properly show the problem.